# optimized linear illumination by 6x in some cases