vectorized the loop over dimensions to calculate all the answers in one go, as per https://github.com/kornia/kornia/issues/2826

LinearCornerIlluminationGenerator :-

Benchmarking with batch_shape=(128, 3, 256, 256), 10 runs...
Original:   25.199 ms/run
Optimized:  4.343 ms/run
Speedup:    5.80×
Max abs diff: 0.000e+00
Mean abs diff: 0.000e+00

LinearIlluminationGenerator:-


Batch=(128, 3, 256, 256), runs=50, device=cuda, dtype=torch.float32, same_on_batch=False
Original : 7.228 ms/run
Optimized: 1.306 ms/run
Speedup  : 5.53×
[gradient] max abs diff: 0.000e+00 | mean abs diff: 0.000e+00 | equal=True

Batch=(128, 3, 256, 256), runs=50, device=cuda, dtype=torch.float32, same_on_batch=True
Original : 6.906 ms/run
Optimized: 1.075 ms/run
Speedup  : 6.42×
[gradient] max abs diff: 0.000e+00 | mean abs diff: 0.000e+00 | equal=True

https://colab.research.google.com/drive/1WFBBO2ENTuyLC1kO8Hm-9clL5ea5GXEN?usp=sharing
heres the benchmarking code for both the functions